+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The changelog contains informations about bug fixes, new features or bc breaking
 
 ### Added
 
+- [#1081](https://github.com/luyadev/luya/issues/1081) Generate Link Interface for internal and external links in order to identify different link types.
 - [#1082](https://github.com/luyadev/luya/issues/1082) Link plugin for ngRest configuration in order to provide internal or external links.
 - [#1003](https://github.com/luyadev/luya/issues/1003) The crawler tag CRAWL_TITLE has been added to ensure a customization of the titles.
 - [#1008](https://github.com/luyadev/luya/issues/1008) Administration interface language can be changed and stored in the user settings.